Improve linter (comments) & remove shared filed. Prefer duplication over worse readability
